// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      enrollments: {
        Row: {
          academic_level: string | null
          course_id: number | null
          enrolled_at: string | null
          enrollment_type: Database["public"]["Enums"]["enrollment_type"]
          grade: number | null
          id: string
          specialty: string | null
          status: Database["public"]["Enums"]["enrollment_status"] | null
          student_id: string | null
        }
        Insert: {
          academic_level?: string | null
          course_id?: number | null
          enrolled_at?: string | null
          enrollment_type?: Database["public"]["Enums"]["enrollment_type"]
          grade?: number | null
          id?: string
          specialty?: string | null
          status?: Database["public"]["Enums"]["enrollment_status"] | null
          student_id?: string | null
        }
        Update: {
          academic_level?: string | null
          course_id?: number | null
          enrolled_at?: string | null
          enrollment_type?: Database["public"]["Enums"]["enrollment_type"]
          grade?: number | null
          id?: string
          specialty?: string | null
          status?: Database["public"]["Enums"]["enrollment_status"] | null
          student_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "enrollments_course_id_fkey"
            columns: ["course_id"]
            isOneToOne: false
            referencedRelation: "courses"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "enrollments_student_id_fkey"
            columns: ["student_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      profiles: {
        Row: {
          academic_degree: string | null
          avatar_url: string | null
          bio: string | null
          created_at: string | null
          date_of_birth: string | null
          department: string | null
          email: string | null
          employee_id: string | null
          full_name: string | null
          headline: string | null
          id: string
          level: string | null
          major: string | null
          mobile: string | null
          role: Database["public"]["Enums"]["app_role"] | null
          specialization: string | null
          student_id: string | null
          subjects: string | null
          updated_at: string | null
          years_of_experience: number | null
        }
        Insert: {
          academic_degree?: string | null
          avatar_url?: string | null
          bio?: string | null
          created_at?: string | null
          date_of_birth?: string | null
          department?: string | null
          email?: string | null
          employee_id?: string | null
          full_name?: string | null
          headline?: string | null
          id: string
          level?: string | null
          major?: string | null
          mobile?: string | null
          role?: Database["public"]["Enums"]["app_role"] | null
          specialization?: string | null
          student_id?: string | null
          subjects?: string | null
          updated_at?: string | null
          years_of_experience?: number | null
        }
        Update: {
          academic_degree?: string | null
          avatar_url?: string | null
          bio?: string | null
          created_at?: string | null
          date_of_birth?: string | null
          department?: string | null
          email?: string | null
          employee_id?: string | null
          full_name?: string | null
          headline?: string | null
          id?: string
          level?: string | null
          major?: string | null
          mobile?: string | null
          role?: Database["public"]["Enums"]["app_role"] | null
          specialization?: string | null
          student_id?: string | null
          subjects?: string | null
          updated_at?: string | null
          years_of_experience?: number | null
        }
        Relationships: []
      }
// ...
    }
// ...
    Enums: {
      app_role: "admin" | "teacher" | "student"
      course_visibility: "active" | "hidden" | "disabled"
      enrollment_status: "enrolled" | "completed" | "dropped"
      enrollment_type: "individual" | "group"
      exam_status: "upcoming" | "ongoing" | "finished"
      material_type: "pdf" | "video" | "link"
      question_type: "mcq" | "true_false" | "essay" | "code"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
  | keyof DefaultSchema["Enums"]
  | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
  ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
  : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
  ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
  : never